requirements.txt
spotipy==2.19.0
python-dotenv==0.19.1